Allow filters to be displayed but inactive
A commonly requested feature is to have the ability to toggle filters on-off in quick succession. This commit achieves such functionality for filters displayed in the header (<Info>) by exposing not only a "remove filter" icon but also a "hide filter" icon. "inactive" filters are not stored in URL state as they're considered temporary. Selected strains and date ranges are not implemented as filters and therefore don't (yet) have this new functionality.

Internally, instead of representing filters as a simple list of values which we are filtering against, we now use a list of objects which allows each item to be "active" or not and allows scope for future functionality. Code is also cleaned up as it was refactored.